wireless-regdb: remove dependency to python attr
Commit 8607edfdb6568 ("wireless-regdb: Parse wmm rule data") introduced
a dependency to the python module attr which is not included by default
in all python installations. Replace the code with manually coding the
constructor instead of using attr. This makes the code also work on
systems without attr.

I would like to avoid an additional dependency in OpenWrt where we
compile the regulatory database inside of the build system.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>