wireless-regdb: restore channel 12 & 13 limitation in the US
The precisely allowed transmission power on channel 12 & 13 is not clear.

I did some fact checking and document reviews in the past, but still
failed to come up with an exact reference or calculation regarding what TX
power should be allowed in general to reduce interference in the
neighboring restricted band 2483.5-2500 MHz and whether indoor vs. outdoor
and b/g/n modulation would make a difference (probably yes).

Existing documents that discuss this are vague and pretty much out of date:
https://en.wikipedia.org/wiki/List_of_WLAN_channels#endnote_B

Following these guidelines, I failed to come up with an approximate
formula that would prohibit usage of channel 12 & 13, but still allow
usage of channel 11 with full power, hence at least part of the argument
for prohibiting channel 12 & 13 needs to be refined.

Until further research, just to stay safe, I propose that we restore the
previous limitation of disabling channel 12 & 13 as before. We should
work towards finding a safe transmission power to use on both and specify
that (like 50mW on channel 13 and 100mW on channel 12?).

I've also started a discussion for investigation on Wikipedia:
https://en.wikipedia.org/wiki/Talk:List_of_WLAN_channels#Should_the_2.4GHz_channel_12_and_13_in_the_USA_need_updating?

Signed-off-by: bkil <b.K.il.h.u+tigbuh@gmail.com>
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>