wireless-regdb: recent FCC report and order allows 5850-5895 immediately
The new band is called U-NII-4.

The report recommends combining it with 5725-5895 to allow 160 MHz
bandwidth, but that's technically not that easy with regdb due to the
differing restrictions of the two parts. Marking the line for U-NII-3
NO-OUTDOOR and PTMP-ONLY along with extending its range would be a
possible workaround, but this needs to be discussed.

I don't see a requirement for TPC, hence reducing EIRP by 3dB is not
needed. I've marked it 33dBm (minus 6dB for clients) to cope with 20MHz,
but the band can support higher power, though the logic is complicated.

The upper subband (5895-5925 MHz) of the new band is reserved for ITS.

"We limit unlicensed use to indoor operations in recognition of the
potential that ITS licensees may currently be operating"

"We also proposed that U-NII-4 devices be permitted to operate at the same
power levels as U-NII-3 devices."

"For the U-NII-4 band, indoor access point EIRP will be limited to
33 dBm/20 MHz and 36 dBm/40 MHz. When combined with U-NII-3 band spectrum,
indoor access point EIRP can scale to 36 dBm for 80 and 160 megahertz
channels."

"Client devices would be limited to power levels 6 dB below the power
limits for access points."

"the First Report and Order prohibit U-NII-4 client-to-client
communications to protect co-channel incumbent ITS"

Signed-off-by: bkil <b.K.il.h.u+tigbuh@gmail.com>
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>