wireless-regdb: Update regulatory rules for Norway (NO) on 6 and 60 GHz
The Norwegian regulatory authority (NKOM) has updated the existing
regulations.

WiFi 6E is now allowed with the same restrictions as in many/most
EU/EEA countries - i.e 200 mW indoor and 25 mW outdoor, Using the
DE entry as a template for now, until this dual entry can be
described in the database.

The update also extended the 60 GHz band to 71 GHz.

The official announcement of the regulation changes:
https://lovdata.no/dokument/LTI/forskrift/2021-06-29-2312

The existing reference in the database has been updated with these
changes, so no new reference is needed.

Signed-off-by: Bjørn Mork <bjorn@mork.no>
Signed-off-by: Seth Forshee <sforshee@kernel.org>