wireless-regdb: Update regulatory rules for South Korea (KR)
This patch is based on MSIT Public Notification 2021-86 ("Unlicensed Radio
Equipment Established Without Notice"), officially announced on 2021-11-29.

The PSD must not exceed 2.5 mW/MHz if the frequency range includes all or
part of 5230-5250 MHz and the bandwidth is equal to or less than 40 MHz.
This leads to the following:
* 5230-5250 @ 20 -> 17 dBm
* 5210-5250 @ 40 -> 20 dBm
Here the power limits for 80/160 MHz bandwidth are also lowered to 17/20 dBm,
as it's not possible to set different power limits for different bandwidths
at the moment.

Extend the last 5 GHz frequency range to 5850 MHz.

WiFi 6E is now allowed with the following restrictions:
* Indoor: the full 1.2 GHz range, up to 160 MHz bandwidth and 2 dBm/MHz PSD
* Outdoor: the lower 500 MHz range, up to 160 MHz bandwidth and 25 mW EIRP
Here only the latter entry is added.

And also update the regulatory source links.

Signed-off-by: Sungbo Eo <mans0n@gorani.run>