wireless-regdb: Update regulatory rules for France (FR) on 6 and 60 GHz
The French national agency for radio frequencies (ANFR) has published an updated
revision of the normative radio band allocation table on 2021-12-14. It includes
the authorization for the lower 6 GHz band for 802.11ax as prescribed by the EU
2021/1067 decision, as well as the authorization for channels 5 and 6 in the 60
GHz band for 802.11ay.

The declaration for the 6 GHz band follows the Low Power Indoor constraints.

https://www.anfr.fr/fileadmin/mediatheque/documents/tnrbf/TNRBF_2021-12-14.pdf

Signed-off-by: Romain Izard <romain.izard.pro@gmail.com>
Signed-off-by: Seth Forshee <sforshee@kernel.org>