wireless-regdb: Update regulatory rules for Israel (IL)
The Israeli Ministry of Communications has published an updated
revision of the normative radio band allocation table on 2021-03-29.
Notable change is the expansion of the range of allowed frequencies
in the 5GHz band.
v2: Increase bandwidth to 160MHz for 5470-5725 MHz range, as there is
no specified bandwidth limits in the document.

Signed-off-by: Roman Kunzman <roman.kunz@tutanota.com>
Signed-off-by: Seth Forshee <sforshee@kernel.org>