wireless-regdb: Unify 6 GHz rules for EU contries
Article 3 of Decision (EU) 2021/1067 states that member states must
adopt rules for harmonised use of the 5945-6425 MHz band by Dec 1, 2021.
Update rules for all member states to include a reference to the
decision, and add the rule for countries which do not yet have it.

Some countries currently have the power limit specified as 200 mW.
Update these countries to use the 23 dBm limit specified by the decision
instead, which is equivalent.

Suggested-by: Dennis Bland <dennis@dbperformance.com>
Signed-off-by: Seth Forshee <sforshee@kernel.org>