wireless-regdb: Update regulatory rules for Russia (RU). Remove DFS requirement.
Use of Wi-Fi devices are regulated by the Decision of the State Commission for Radio Frequencies
 under the Ministry of Communications of Russia of February 29, 2016 No. 16-36-03
 “On Amendments to the Decision of the State Committee for Radio Frequencies of May 7, 2007 No. 07-20-03-001
 “On the allocation of radio frequency bands to short-range devices »

https://www.garant.ru/products/ipo/prime/doc/71249926/

The document sets 100mW for 2.4 GHz and 200 mW for 5 GHz.
There is a special limitation 100mW for 5 GHz onboard aircraft.

DFS was never required in Russia.
Instead the authority chose to prohibit outdoor use for 5 GHz totally.

There is a vague comment that power should be automatically reduced in direction of other devices.
This can be interpreted as a TPC requirement, but in practice devices are certified with 200mW.

Signed-off-by: Dmitry Tunin <hanipouspilot@gmail.com>
[ saf: keep 5 GHz power limits at 100 mW due to TPC requirement ]
Signed-off-by: Seth Forshee <sforshee@kernel.org>